functions_py/requirements.txt
firebase-functions>=0.4.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
